requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.1
Django==3.1.4
et-xmlfile==1.1.0
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
openpyxl==3.0.9
Pillow==9.1.0
PyPDF2==1.27.12
python-bidi==0.4.2
pytz==2022.1
reportlab==3.6.6
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
webencodings==0.5.1
whitenoise==6.1.0
xhtml2pdf==0.2.5
python-dotenv==1.0.0
mysqlclient